Pulls the core ExecutionTests back into scalding-core
Leave Means into commons for its extra dep needs
